requirements.txt
twilio
python-dotenv
deep_translator
langdetect